==> Geographical Infrastructure/Codes_2023/01_PropertiesToParcel-YET.Rmd <==
---
title: "01_PropertiesToParcel_2023-07_YET"
output: html_document
date: "2023-07-18"
---

# 0. Preliminary Stuff

```{r Set Environment(optional)}
## If you use the drive file format, you can use the code below. If not, put your path here.
GDrivepath <- "~/Google Drive/BARI Research Team Data Library/"
setwd(paste0(GDrivepath,"Geographical Infrastructure/Boston Geographical Infrastructure 2019/Scripts"))
```

```{r Load Libraries}
library(compare)
library(stringr)
library(tidyverse)
```

?????????? ### Determining the same properties with dissimilar GIS_ID ----------------------#

```{r Load Data}
##Download the raw cross-sectional property assessment and land parcels files from Analyze Boston
propassraw <- read.csv("~/Desktop/BARI/GI-2023/new datasets/Property Assesments/fy2023-property-assessment-data.csv")
# downloaded from AB: https://data.boston.gov/dataset/property-assessment
#parcelsraw <- read.csv("~/Desktop/BARI/GI-2023/new datasets/Parcel Datasets/Parcels_2023.csv")
# downloaded from AB: https://data.boston.gov/dataset/parcels-20181
```

From ##SSH

# 1. Data Manipulation

????? # 1. Determine without doubt that GIS_ID and PID_LONG (or the most current Parcel ID in the parcels file) refer to the same unit

## Updates on column names:
-  August 2021: the column ST_NAME_SUF does not exist anymore;the PTYPE column changed name to LUC
-  July 2023: the column ZIPCODE changed name to ZIP_CODE

```{r Subsetting}
prop23 <- subset(propassraw, select = c("PID","CM_ID","GIS_ID","ST_NUM",
                                        "ST_NAME","UNIT_NUM","ZIP_CODE","OWNER",
                                        "LUC", "LU","LAND_SF",
                                        "YR_BUILT")) 

# 180627 observations
```


```{r Remove Dataset}
rm(propassraw)
```


## 1.a Create vectors of just GIS_ID and PID_LONG

YET: As of August 23rd, there is no longer a PID_LONG column. Instead, there is a MAP_PAR_ID column.

```{ Create Vectors}
gisid <- prop23$GIS_ID[!duplicated(prop23$GIS_ID)]
#pidlong <- parcelsraw$MAP_PAR_ID[!duplicated(parcelsraw$MAP_PAR_ID)] 
```

## 1.b   Determine the difference between them visually

```{r}
#View(table(setdiff(gisid, pidlong)))
#prop23_diff <- subset(prop23, prop23$GIS_ID %in% c(setdiff(gisid, pidlong)))
```


## 1.c Remove Repeated Addresess

SSH - This part is to remove repeated addresses 
  - if !duplicated right now, nrow = 172477
  - Cleaning the street numbers to separate the range of street numbers.
  - This is required for expanding individual addresses

```{r Creating a reference ID}
# It enables us to see the change in the range of numbers after as.numeric()
prop23$FullID <- paste(prop23$PID, prop23$ST_NUM, toupper(prop23$ST_NAME),
                       prop23$ZIP_CODE) #column name has changed in 2023
                       #prop23$ST_NAME_SUF, does not exist anymore (Aug23)
                       #prop23$GIS_ID)

prop23$fulladress <- paste(prop23$ST_NUM, toupper(prop23$ST_NAME),
                       prop23$ZIP_CODE)

prop23$ST_NAME <- toupper(prop23$ST_NAME)
```



SSH: not separating the ranges at this point
YET: ??? if we separate the ranges, could not use st_num
```{ Separate the Ranges}
prop23$street_num <- prop23$ST_NUM
#new
prop23 <- separate(prop23, ST_NUM, into = c("num1","num2"), sep = " ")
```

## 1.d Reformat the Dataset

We need to reformat the dataset for an easier manipulation.

```{ Reformat the Dataset}
prop23_sub <- prop23 %>%
    mutate(pid = PID,
           #num1 = num1,
           #num2 = num2,
           street_num = ST_NUM,
           street_c = ST_NAME,
           #suffix_c = ST_NAME_SUF, #this does not exist in 2021
           zip_c = ZIP_CODE,
           GIS_ID = GIS_ID) %>%
           #select(FullID, pid, num1, num2, street_c, suffix_c, zip_c, GIS_ID) # 177091
           select(FullID, pid, street_num, street_c, zip_c, GIS_ID)
```


## 1.e Manipulate street_num column


```{ Separating "street_num" column}
Let's separate street_num ranges into rows

prop23_sub$street_num <- gsub("[^0-9\\.]", " ", prop23_sub$street_num) 

prop23_adds <- separate_rows(prop23_sub, street_num, convert = TRUE) # 212872

prop23_adds$id <- paste(prop23_adds$pid,prop23_adds$street_num,
                        prop23_adds$street_c,prop23_adds$suffix_c,
                        prop23_adds$zip_c,
                        prop23_adds$GIS_ID,
                        sep = "+")
 
sum(duplicated(prop23_adds$id)) # 16402
 
prop23_adds <- prop23_adds[!duplicated(prop23_adds$id),] # 196470

prop23_adds$FullID <- paste(prop23_adds$pid,
                            prop23_adds$street_c,prop23_adds$suffix_c,prop23_adds$zip_c,
                            prop23_adds$GIS_ID)
 
prop23_adds$flag <- ifelse(is.na(prop23_adds$street_num) & duplicated(prop23_adds$FullID), 0, 1)

prop23_adds <- prop23_adds %>% filter(flag == 1) # 191914
```

SSH: this is a properties file with the information we have on the street numbers without full expansion 
-- skip the rest



We need to convert street num1 into a numeric for future use.

```{ Convert street num1 into a numeric form}
prop23_sub$num1 <- as.numeric(prop23_sub$num1)
```


We also need to change the num2 address to equal num1 if num1 was NA. 
This makes it easier to identify individual addresses.


```{ Change num2}
prop23_sub$num2 <- ifelse(is.na(prop23_sub$num2),prop23_sub$num1, as.numeric(prop23_sub$num2))

```



Then, we need to create a unique id for each row (which is each variable in the entire row pasted together)
```{ Create unique ids}
prop23_sub$id <- paste(prop23_sub$pid,prop23_sub$num1,prop23_sub$num2,
                       prop23_sub$street_c,prop23_sub$zip_c,
                       prop23_sub$GIS_ID,
                       sep = "+") # prop23_sub$suffix_c was removed in 2021.
```

Then, we need to get rid of duplicates.
```{r Remove duplicates}
prop23_sub <- prop23_sub[!duplicated(prop23_sub$FullID),]
```

Report the nrow here: 
```{r}
nrow(prop23_sub)
```

- SSH - nrow was 171922 before.
- YET - nrow is 180448 now.

- SSH 
  - we could not do the expansion anymore due to possible duplication but 
  - we still need to separate unit ranges from addresses 


Some of the address numbers range from high to low (e.g. 4-2 rather than 2-4). 
This switches the range back from low to high


```{ Arrange num columns}
prop23_sub <- prop23_sub %>%
    group_by(num1>num2) %>%
    mutate(to=num1,
           from=num2) %>%
    mutate(num1=from,
           num2=to) %>%
    ungroup() %>%
    select(pid, num1, num2, street_c, zip_c, GIS_ID, FullID)
```

# 2. Data Generation
The following code creates a new dataframe from the single (e.g. 1-2) ranges of addresses (all others are assumed to be from an odd or even range)

```{ Create a new dataframe from single ranges}
prop23_one <- prop23_sub %>%
    filter(num2 == num1+1)%>%
    rowwise()%>%
    do(data.frame(FullID=.$FullID, pid=.$pid, addressnum=seq(.$num1, .$num2, by=1),
                  street_c=.$street_c, zip_c=.$zip_c,
                  GIS_ID=.$GIS_ID)) # suffix_c=.$suffix_c removed 2021
```


The following code creates a new dataframe from the divisible by 2 (e.g. 2-4) range of addresses (where each address is assumed to be on one side of the street at odd or even intervals)
```{ Create a new dataframe the divisible by 2}
prop23_two <- prop23_sub %>%
    filter(num1 < num2 & num1 != num2+1 & num2 - num1 < 21)%>%
    rowwise()%>%
    do(data.frame(FullID=.$FullID, pid=.$pid, addressnum=seq(.$num1, .$num2, by=2),
                  street_c=.$street_c, zip_c=.$zip_c,
                  GIS_ID=.$GIS_ID))  # suffix_c=.$suffix_c removed in 2021
```
 


The following code creates a new dataframe from the individual addresses without range
```{ Create a new dataframe}
prop23_three <- prop23_sub %>%
    filter(c(num1 == num2 | (num1>=0 & is.na(num2))))%>%
    mutate(addressnum=.$num1)%>%
    select(FullID,pid,addressnum,street_c,zip_c,GIS_ID) #suffix_c, removed in 2021
```




The following code creates a new dataframe  for addresses on with a num2
```{ Create a new dataframe}
prop23_four <- prop23_sub %>%
    filter(c(is.na(num1) & num2 >= 0)) %>%
    mutate(addressnum = .$num2) %>%
    select(FullID,pid,addressnum,street_c,zip_c,GIS_ID) #suffix_c, removed in 2021

```

The following code creates a new dataframe from the NA addresses
```{ Create a new dataframe}
prop23_five <- prop23_sub %>%
    filter(c(is.na(num1) & is.na(num2))) %>%
    mutate(addressnum= "-") %>%
    select(FullID,pid,addressnum,street_c,zip_c,GIS_ID) #suffix_c, removed in 2021
```


The following code creates a new dataframe for the addresses numbers that APPARENTLY won't switch
```{ Create a new dataframe}
prop23_six <- prop23_sub %>%
    filter(num1+1==num2) %>%
    rowwise()%>%
    do(data.frame(FullID=.$FullID, pid=.$pid, addressnum=seq(.$num1, .$num2, by=1),
                  street_c=.$street_c, zip_c=.$zip_c,
                  GIS_ID=.$GIS_ID)) # suffix_c=.$suffix_c, removed in 2021
```


The following code creates a new dataframe for the addresses numbers that wouldn't switch and range by 2
```{ Create a new dataframe}
prop23_seven <- prop23_sub %>%
    filter(c((num1>num2) & num2 != (num1+1) & num1 - num2 < 21)) %>%
    rowwise()%>%
    do(data.frame(FullID=.$FullID, pid=.$pid, addressnum=seq(.$num2, .$num1, by=2),
                  street_c=.$street_c, zip_c=.$zip_c,
                  GIS_ID=.$GIS_ID)) # suffix_c=.$suffix_c removed in 2021
```


The following code creates a new dataframe for num2 - num1 > 21
```{ Create a new dataframe - num2 - num1 > 21}
prop23_eight <- prop23_sub %>%
    filter(abs(num1 - num2) > 21) %>%
    rowwise()%>%
    do(data.frame(FullID=.$FullID, pid=.$pid, addressnum=paste(.$num2, .$num1, sep = "-"),
                  street_c=.$street_c, zip_c=.$zip_c,
                  GIS_ID=.$GIS_ID)) # suffix_c=.$suffix_c removed in 2021
```


# 3. Data Combination

We will adding all of the dataframes together and creating unique IDs for Land_Parcel_ID conversion

```{r}
#prop23_adds <- rbind(prop23_one,prop23_two,prop23_three,prop23_four,prop23_five,
                     #prop23_six, prop23_seven, prop23_eight) # 180448

prop23_adds <- prop23_sub
#prop23_adds <- rename(prop23_adds, adressnum = street_num)

```



# 4. Checking  
The code below is to make sure all GIS_IDs transfered from prop23_sub to prop23_adds.
```{ Checking the transfer}
View(table(setdiff(prop23_sub$GIS_ID, prop23_adds$GIS_ID))) # this should be empty
```

## 4.a. Add fulladdress variable
We need to add fulladdress variable for checks and fulladdress+pid to rid duplicates (we did this before!!!!!!!)

```{r Adding fulladdress variable}
#prop23_adds$fulladdress <- paste(prop23_adds$addressnum,prop23_adds$street_c,
#                                 prop23_adds$zip_c) #prop23_adds$suffix_c removed 2021
#prop23_adds$gisaddress <- paste(prop23_adds$addressnum,prop23_adds$street_c,
#                                prop23_adds$zip_c,
#                                prop23_adds$GIS_ID) #prop23_adds$suffix_c, removed 2021
#prop23_adds$fulladdressp <- paste(prop23_adds$addressnum,prop23_adds$street_c,
#                                  prop23_adds$zip_c,
#                                  prop23_adds$pid) #prop23_adds$suffix_c removed 2021
prop23 <- prop23[!duplicated(prop23$FullID),]
prop23$Land_Parcel_ID <- NA
prop23$nastreet <- ifelse(is.na(prop23$ST_NUM) == TRUE, 1, 0) # 176925
```

## 4.b. Creating a vector of duplicated fulladdress

We need to create a vector of duplicated fulladdress since duplicated() will negate the first match

```{r First way}
#duplicates <- prop23_adds %>%
#                     filter(duplicated(fulladdress) & nastreet==0) %>%
#                     select("fulladdress")
```


```{r Second way}
duplicates <- 
    prop23[which(duplicated(prop23$fulladress) &  prop23$nastreet==0),] #77952 in Aug 23 # 74207
# duplicates <- prop23_adds[which(duplicated(prop23_adds$fulladdress)), "fulladdress"] # 83475
```



## 4.c. Making a dummy-variable of duplicated fulladdresses

```{r Create a dummy variable}
prop23$dupadd = ifelse(prop23$fulladress %in% duplicates$fulladress, 1, 0)
prop23$Land_Parcel_ID = ifelse(prop23$dupadd==0, prop23$GIS_ID, NA) # 176925
# prop23_adds <- prop23_adds %>%
#     mutate(dupadd = ifelse(fulladdress %in% duplicates$fulladdress, 1, 0),
#            Land_Parcel_ID = ifelse(dupadd==0, GIS_ID, NA)) 
```


## 4.d. Find the minimum shared GIS_ID

We need to aggregate to find the minimum shared GIS_ID among duplicated fulladdress

```{r Aggregation}
mingis <- aggregate(GIS_ID~fulladress, prop23[prop23$dupadd==1,],
                    FUN = min)
```

### 4.d.1. Add a dummy

```{r Adding a dummy}
# Creating a dummy-variable for the minimum GIS_ID
mingis$lowgis <- 1
```

## 4.e. Merge dataframes
                                                
At the end, we need to merge the minimum GIS_ID dataframe onto the full dataframe and clean the final dataframe

```{r Join dataframes}
prop23 <- left_join(prop23, mingis, by=c("fulladress")) #180,448 in Aug 23 # 176925

prop23 <- prop23 %>%
    group_by(lowgis==1) %>%
    mutate(Land_Parcel_ID = ifelse(!is.na(GIS_ID.y), GIS_ID.y, GIS_ID.x)) %>%
    ungroup() %>%
    mutate(GIS_ID=GIS_ID.x) %>%
    select(PID, ST_NUM, ST_NAME, ZIP_CODE, fulladress, GIS_ID, Land_Parcel_ID) 
        # 176,925 #suffix_c removed 2021
```


    
```{r Print csv}
write_csv(prop23, "~/Desktop/BARI/GI-2023/outputs/PropertiesToParcels-091423_YET.csv")
#write.csv(prop23_adds, "PropertiesToParcels20230329.csv", row.names = F)
```


# 5. Sanity Checks

???? SSH
```{r}
# some sanity checks
prop23 %>%
  group_by(Land_Parcel_ID, GIS_ID) %>%
  tally() %>%
  arrange(desc(n))

prop23 %>%
   group_by(Land_Parcel_ID) %>%
   summarise(n_distinct(GIS_ID)) %>%
   View()
```
